Dynamic loading of font data and texture
Font json and png data is loaded by fetch() calls rather than from
a massive inlined JSON file, which greatly reduces the webpack
output size and also makes compiling/editing much more responsive.

Create-react-app's webpack.config.js is monkey patched during script
calls to insert the necessary settings for packing font resources.

TextEditor reworked so all the functions that only work once a font is
loaded have a strong type binding on 'this' rather than having to do
lots of this!.font or if(this.font) everywhere. It's a touch clunky around
the event handlers but seems to be an improvement.